# src/orders/openapi.yml
openapi: 3.0.3
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /orders:
    get:
      tags:
        - orders
      summary: List orders
      operationId: listOrders
      parameters:
        - name: customerEmail
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExistingOrder"
    post:
      tags:
        - orders
      summary: Create an order
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExistingOrder"
  /users:
    post:
      tags:
        - users
      summary: Register a user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/login:
    post:
      tags:
        - authentication
      summary: Validate user credentials
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
components:
  schemas:
    UserRegistration:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    Credentials:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
    Order:
      type: object
      properties:
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
          format: int32
        totalCost:
          type: integer
          format: int32
        unitCost:
          type: integer
          format: int32
    ShippingAddress:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        address1:
          type: string
        address2:
          type: string
        city:
          type: string
        zipCode:
          type: string
        state:
          type: string
    ExistingOrder:
      type: object
      properties:
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        id:
          type: string
        createdDate:
          type: string
          format: date-time